+++ lib/pages/player/danmaku_destination_sheet.dart
@@ -3,11 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:kazumi/bean/dialog/adaptive_bottom_sheet.dart';
 import 'package:kazumi/bean/dialog/material_bottom_sheet.dart';
 import 'package:kazumi/bean/widget/split_list_row.dart';
-
-enum DanmakuDestination {
-  chatRoom,
-  remoteDanmaku,
-}
+import 'package:kazumi/pages/player/controller/player_models.dart';
 
 Future<DanmakuDestination?> showDanmakuDestinationSheet(BuildContext context) {
   return showAdaptiveBottomSheet<DanmakuDestination>(